refactor: BaseEntity를 BaseEntity와 BaseUpdatableEntity로 분리, entity에서 toString, equals 등 제거
