--- src/main/java/com/sprint/mission/discodeit/entity/BaseEntity.java
@@ -1,35 +1,0 @@
-package com.sprint.mission.discodeit.entity;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import java.io.Serial;
-import java.io.Serializable;
-import java.time.Instant;
-import java.util.UUID;
-
-@Getter
-public abstract class BaseEntity implements Serializable {
-    @Serial
-    private static final long serialVersionUID = 1L;
-    private UUID id;
-    private Instant createdAt;
-
-    @Setter
-    private Instant updatedAt;
-
-    public BaseEntity() {
-        this.id = UUID.randomUUID();
-        this.createdAt = Instant.now();
-        this.updatedAt = Instant.now();
-    }
-
-    @Override
-    public String toString() {
-        return "BaseEntity{" +
-                "id=" + id +
-                ", createdAt=" + createdAt +
-                ", updatedAt=" + updatedAt +
-                '}';
-    }
-}
