+++ src/main/java/com/sprint/mission/discodeit/entity/BinaryContent.java
@@ -1,42 +1,19 @@
 package com.sprint.mission.discodeit.entity;
 
+import com.sprint.mission.discodeit.entity.base.BaseEntity;
 import lombok.Getter;
 
-import java.io.Serial;
-import java.io.Serializable;
-import java.time.Instant;
-import java.util.Arrays;
-import java.util.UUID;
-
 @Getter
-public class BinaryContent implements Serializable {
-    @Serial
-    private static final long serialVersionUID = 1L;
-    private UUID id;
-    private Instant createdAt;
+public class BinaryContent extends BaseEntity {
     private String fileName;
     private Long size;
     private String contentType;
     private byte[] bytes;
 
     public BinaryContent(String fileName, Long size, String contentType, byte[] bytes) {
-        this.id = UUID.randomUUID();
-        this.createdAt = Instant.now();
         this.fileName = fileName;
         this.size = size;
         this.contentType = contentType;
         this.bytes = bytes;
-    }
-
-    @Override
-    public String toString() {
-        return "BinaryContent{" +
-                "id=" + id +
-                ", createdAt=" + createdAt +
-                ", fileName='" + fileName + '\'' +
-                ", size=" + size +
-                ", contentType='" + contentType + '\'' +
-                ", bytes=" + Arrays.toString(bytes) +
-                '}';
     }
 }
