+++ src/main/java/com/sprint/mission/discodeit/entity/ReadStatus.java
@@ -1,16 +1,13 @@
 package com.sprint.mission.discodeit.entity;
 
+import com.sprint.mission.discodeit.entity.base.BaseUpdatableEntity;
 import lombok.Getter;
 
-import java.io.Serial;
-import java.io.Serializable;
 import java.time.Instant;
 import java.util.UUID;
 
 @Getter
-public class ReadStatus extends BaseEntity implements Serializable {
-    @Serial
-    private static final long serialVersionUID = 1L;
+public class ReadStatus extends BaseUpdatableEntity {
     private UUID userId;
     private UUID channelId;
     private Instant lastReadAt;
