+++ src/main/java/com/sprint/mission/discodeit/entity/base/BaseEntity.java
@@ -1,0 +1,17 @@
+package com.sprint.mission.discodeit.entity.base;
+
+import lombok.Getter;
+
+import java.time.Instant;
+import java.util.UUID;
+
+@Getter
+public abstract class BaseEntity {
+    private UUID id;
+    private Instant createdAt;
+
+    public BaseEntity() {
+        this.id = UUID.randomUUID();
+        this.createdAt = Instant.now();
+    }
+}
